Cherry-pick change from 0482ab6, so that v11.1.2 build passes
The code added in this commit was removed in f860685aa, while a PR merged
around the same time used this message key. 0482ab6 readded the message+key,
to get unit tests passing, even though the intent of the commit was larger.

This commit cherry-picks just the needed change from 0482ab6.