Await account render after unlock before checking for mocked endpoint resolution in unlock-wallet.spec.js (#21422)
Fixes the flakiness of the `should send first three Page metric events
upon fullscreen page load` test when run as part of the `
test-e2e-chrome-mmi ` job. While this seems to fix the issue, I don't
have a 100% definite explanation. My hypothesis in making the change in
this PR is that occasionaly on mmi builds the e2e tests would take
longer to get past the unlock stage than expected, and so it is possible
that not all metrics requests that the `unlock-wallet.spec.js` test is
waiiting for have be made by the time the test starts looking for them.

Anecdotal evidence seems to suggest that this has worked, and that this
test is passing.