Fix flakiness of import-flow.spec.js and swaps-notifications.spec.js (#21464)
## **Description**
This PR fixes two outstanding flaky tests.

The first is swaps-notifications.test.js. An example of its failure is
https://app.circleci.com/pipelines/github/MetaMask/metamask-extension/60381/workflows/e8302bb0-0e22-4316-b322-ec33196994c8/jobs/1879141/tests

This test is sometimes failing because the `checkNotification` function
was getting `boxContent` regardless of what text is included, and it is
possible that at the time of getting `boxContent` the component has not
finished calculating the data, and so the subsequent assertion fails.

The below video is a slowed down screen capture of the e2e test, and it
shows how the content of the box is briefly different than expected.


https://github.com/MetaMask/metamask-extension/assets/7499938/e3245643-d792-424e-9a18-927282f1730c

The fix is to use the expected text as part of the locator to find the
element. This was also applied to the title element that is found within
`checkNotification`

The second is is `import-flow.spec.js`. The failures here were happening
because the tests were attempting to click the account menu button
immediately after action was taken to close a separate modal, and the
partially opaque modal background was interferring with the tests
attempt to click the button. The fix is to use `findClickableElement`
instead of `findElement`, so that we know the element is truly ready and
able to be clicked before the test attempts to click it.


## **Manual testing steps**

all tests should pass

## **Pre-merge author checklist**

- [x] I’ve followed [MetaMask Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've clearly explained:
  - [ ] What problem this PR is solving.
  - [ ] How this problem was solved.
  - [ ] How reviewers can test my changes.
- [x] I’ve indicated what issue this PR is linked to: Fixes #???
- [x] I’ve included tests if applicable.
- [ ] I’ve documented any added code.
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/develop/.github/guidelines/LABELING_GUIDELINES.md)).
- [ ] I’ve properly set the pull request status:
  - [ ] In case it's not yet "ready for review", I've set it to "draft".
- [ ] In case it's "ready for review", I've changed it from "draft" to
"non-draft".

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.