e2e: fixing automatic installation of webdriver (#21707)
Before this PR, to run E2E locally, most people had to run this or similar
`node_modules/selenium-webdriver/bin/macos/selenium-manager --browser chrome`

But actually Selenium Manager is **supposed** to be automatic, and just transparently work.

After some investigation, I determined the problem to be the proxy.  Selenium Manager was trying to download the drivers through the proxy, and it wasn't working.

I changed the format of the proxy from `options.setProxy` to individual Chrome and Firefox options, and it works now.