fix: skip flaky errors.spec test (#21761)
The end-to-end test that was originally flaky doesn't send sentry events
for all migrations, only a subset of them.

I am still not 100% sure about the original source of the flakiness. One
possible explanation is that migrations take a variable amount of time
to run.

This PR skips the flaky test until we understand more about the root
cause.